src/closed_models/requirements.txt
openai
pyyaml
anthropic